Move _types.image to types.image
