+++ towhee/hub/builtin/operators/computer_vision.py
@@ -16,8 +16,8 @@ from pathlib import Path
 import numpy as np
 
 from towhee.engine import register
-from towhee._types.image import Image
-from towhee.types.image_utils import to_image_color
+from towhee.types.image import Image
+from towhee.utils.image_utils import to_image_color
 
 # pylint: disable=import-outside-toplevel
 # pylint: disable=invalid-name
